simplify defining the initial_sort
